requirements.txt
annotated-types==0.7.0
anthropic==0.94.0
anyio==4.13.0
asgiref==3.9.1
certifi==2025.8.3
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.3
cloudinary==1.44.1
cryptography==46.0.7
decouple==0.0.7
distro==1.9.0
dj-database-url==3.0.1
Django==5.2
django-appconf==1.2.0
django-channels==0.7.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-cryptography==1.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
docopt==0.6.2
docstring_parser==0.17.0
gitignore==0.0.8
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.14.0
msgpack==1.1.1
oauthlib==3.3.1
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.9.0
pypng==0.20220715.0
PyQRCode==1.2.1
python-dotenv==1.0.1
python-magic==0.4.27
pytube==15.0.0
redis==6.4.0
requests==2.32.5
requests-oauthlib==2.0.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tailwindcss==0.0.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wheel==0.45.1
whitenoise==6.9.0
